refactor vscode folder store
