+++ src/client/src/containers/RightSidebar/About/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useSelector } from "react-redux";
 import styles from "./styles.module.css";
-import { getVersionsSelector } from "../../../store/vscode/vscodeApiSelector";
+import { getVersionsSelector } from "../../../store/vscode/selector";
 import { IVersions } from "../../../types/version";
 import { InjectedIntlProps, injectIntl } from "react-intl";
 import { AppState } from "../../../store/combineReducers";
