--- src/client/src/store/vscode/VSCodeActionType.ts
@@ -1,5 +1,0 @@
-import { IVSCodeAPIActionType } from "./model";
-
-type VSCodeActionTypes = IVSCodeAPIActionType;
-
-export default VSCodeActionTypes;
