+++ src/client/src/store/vscode/reducer.ts
